Set single cource of truth for prices and ensure that keyboard actions don't cause border in Checkout page

+++ CheckoutKata.Core/ICheckout.cs
@@ -4,12 +4,13 @@ public interface ICheckout
 {
     void Scan(string item);
 
-    int GetTotalPrice();
+    decimal GetTotalPrice();
 
     IReadOnlyList<string> ScannedItems { get; }
 
     void RemoveLast();
 
     void ClearAll();
+
     void RemoveOne(string sku);
 }
